chore: add listing, booking, amenity models

+++ prisma/seed/seed.ts
@@ -1,8 +1,15 @@
 import { prisma } from '../prisma-client'
 
+import { createAmenities } from './create-amenities'
+import { createBookingDays } from './create-booking-days'
+import { createBookings } from './create-bookings'
+import { createListingUnits } from './create-listing-units'
+import { createListings } from './create-listings'
 import { createProfiles } from './create-profiles'
+import { createReviews } from './create-reviews'
 import { createRoles } from './create-roles'
 import { createTokens } from './create-tokens'
+import { createUnitCalendarDays } from './create-unit-calendar-days'
 import { createUsers } from './create-users'
 
 const up = async () => {
@@ -13,6 +20,22 @@ const up = async () => {
 
   await createProfiles(userIds)
   await createTokens(userIds)
+
+  await createAmenities()
+
+  const createdListings = await createListings(createdUsers)
+  const createdListingUnits = await createListingUnits(createdListings)
+
+  await createUnitCalendarDays(createdListingUnits)
+
+  const createdBookings = await createBookings(
+    createdUsers,
+    createdListingUnits,
+    createdListings
+  )
+
+  await createBookingDays(createdBookings)
+  await createReviews(createdBookings)
 }
 
 const down = async () => {
@@ -21,6 +44,10 @@ const down = async () => {
   >`SELECT tablename FROM pg_tables WHERE schemaname='public'`
 
   for (const { tablename } of tables) {
+    if (tablename === '_prisma_migrations') {
+      continue
+    }
+
     await prisma.$executeRawUnsafe(
       `TRUNCATE TABLE "${tablename}" RESTART IDENTITY CASCADE`
     )
